fly global declaration
